Moved loading of related songs for a song to the Song class (obviously)
